requirements.txt
aioredis==1.3.1
amqp==2.6.1
asgiref==3.2.10
async-timeout==4.0.0a2
atomicwrites==1.4.0
attrs==20.2.0
autobahn==20.7.1
Automat==20.2.0
beautifulsoup4==4.9.3
billiard==3.6.3.0
bs4==0.0.1
celery==4.4.7
certifi==2020.12.5
cffi==1.14.3
channels==3.0.1
channels-redis==3.2.0
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
colorama==0.4.4
constantly==15.1.0
cryptography==3.1.1
defusedxml==0.7.0rc2
Django==3.1.2
django-autoslug==1.9.8
django-braces==1.14.0
django-celery-beat==2.1.0
django-celery-results==1.2.1
django-ckeditor==6.0.0
django-common-helpers==0.9.2
django-cors-headers==3.5.0
django-crontab==0.7.1
django-extensions==3.1.0
django-filter==2.4.0
django-js-asset==1.2.2
django-mailer==2.1
django-model-search==2.2
django-model-utils==4.0.0
django-modeltranslation==0.16.2
django-mptt==0.11.0
django-multiselectfield==0.1.12
django-oauth-toolkit==1.3.3
django-rest-framework-social-oauth2==1.1.0
django-templated-mail==1.1.1
django-timezone-field==4.0
django-versatileimagefield==2.0
django-websocket-request==0.1.1
djangochannelsrestframework==0.1.1
djangorestframework==3.12.1
djangorestframework-simplejwt==4.4.0
djoser==2.0.5
hiredis==1.1.0
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
inflection==0.5.1
iniconfig==1.1.1
kombu==4.6.11
lockfile==0.12.2
lxml==4.6.2

mongoengine==0.22.1
msgpack==1.0.0
ndg-httpsclient==0.5.1
oauthlib==3.1.0
packaging==20.4
pathlib==1.0.1
Pillow==8.0.1
pluggy==0.13.1
prompt-toolkit==3.0.8
py==1.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
PyJWT==2.0.1
pymongo==3.11.2
PyMySQL==0.10.1
pyOpenSSL==19.1.0
pyparsing==2.4.7
pytest==6.1.1
python-crontab==2.5.1
python-dateutil==2.8.1
python-magic==0.4.15
python-slugify==4.0.1
python3-openid==3.2.0
pytils==0.3
pytz==2020.1
redis==3.5.3
requests==2.25.1
requests-oauthlib==1.3.0
service-identity==18.1.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.0.3
social-auth-storage-mongoengine==1.0.1
soupsieve==2.2
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.1
Twisted==20.3.0
txaio==20.4.1
typing-extensions==3.7.4.3
urllib3==1.26.3
vine==1.3.0
wcwidth==0.2.5
zope.interface==5.1.2
python-gettext==4.0
gunicorn
psycopg2
daphne
django-cachalot
psycopg2